refactor(audio): extract probe/split/converter/detect-format from AudioProcessor (move-only)
